refactor: add some folder structure
